requirements.txt
amqp==2.6.1
asgiref @ file:///tmp/build/80754af9/asgiref_1594338739818/work
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.3
django-allauth==0.42.0
django-extensions==3.0.5
django-filter==2.3.0
django-redis==4.12.1
django-rest-auth==0.9.5
djangorestframework==3.11.1
drf-access-policy==0.7.0
fcm-django==0.3.4
idna==2.10
kombu==4.6.11
Markdown==3.2.2
oauthlib==3.1.0
pyfcm==1.4.7
python3-openid==3.2.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
vine==1.3.0
wincertstore==0.2
